Prevent adding duplicate mint infos
When we add mint infos, we must specify how many mint infos we have already configured.

We can use this to avoid adding duplicate mint infos. The most straightforward way to add 10k mint infos is to add 100 at a time. Then every time when we add another batch `b` of 100, we say that we expect to have added `b * 100` of mint infos already. If this instruction is retried in a different transaction, it will fail.